requirements.txt
# we're not compatible with the latest nova API
python-novaclient==7.1.0
PyYAML==3.12
jinja2==2.9.6
awscli==1.11.72
pykwalify==1.6.0
boto3==1.4.4
docker==2.4.2
paramiko==2.2.1
